feat: estruturando editar projeto
