+++ src/app/_helpers/svg/closeIcon.tsx
@@ -13,7 +13,7 @@ export function CloseIcon({
       viewBox="0 0 24 24"
       fill="none"
     >
-      <g clip-path="url(#clip0_529_10366)">
+      <g clipPath="url(#clip0_529_10366)">
         <path
           d="M19 6.41L17.59 5L12 10.59L6.41 5L5 6.41L10.59 12L5 17.59L6.41 19L12 13.41L17.59 19L19 17.59L13.41 12L19 6.41Z"
           fill="#323232"
